+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Downloads/API**) Add batch GQL mutation for reordering chapter downloads (`reorderChapterDownloads`)
 - (**Logs**) Add IP location logging
 - (**Cache/API**) Add a way to clear cache & cookies
+- (**API**) Add platform information to `aboutServer` query
 
 ### Changed
 - (**SystemTray**) Disable DorkBox update requests